Remove dependency on causes.git registration
The specs depended on a base registry being set up, which happens in
causes.git in a config/initializer/ file. In the absence of a registry,
we were exploding on trying to restore the old (`nil`) registry after
each example, which caused the `#validate!` method to complain loudly
(as designed).

This small change to the spec suite shields us from that detail.

Change-Id: Ife68a58e9aa9efdad5794dd7b527b0e485f789ad
Reviewed-on: https://gerrit.causes.com/26275
Reviewed-by: Lann Martin <lann@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>