Allow PubSubHub to accept unregistered events
Currently, if you send an event to PubSubHub to which nothing is
subscribed, an error is raised. This seems counter to the implied
separation between publishers and subscribers - the publishing of any
piece of content should succeed regardless of the presence of
subscribers.

This adds a simple fix to allow a sensible default behavior when nothing
is registered for an event.

Change-Id: I9188de1573c072c4708b8ea2178d17f7e0c7c61f